incercare2

+++ src/main/java/com/oana/proiect/controller/CustomerController.java
@@ -21,7 +21,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}")
-    public Customer getCustomerById(@PathVariable Long id) {
+    public Customer getCustomerById(@PathVariable Integer id) {
         return customerService.getCustomerById(id);   // returneaza un client dupa id 
     }
